Move common OpenInZeplin files
