+++ src/analytics/Analytics.ts
@@ -1,7 +1,7 @@
 import MixpanelHelper from "./util/MixpanelHelper";
 import BarrelType from "../common/domain/barrel/BarrelType";
 import PinType from "../sidebar/pin/model/PinType";
-import ZeplinLinkType from "../sidebar/openInZeplin/model/ZeplinLinkType";
+import ZeplinLinkType from "../common/domain/openInZeplin/model/ZeplinLinkType";
 import JiraType from "../common/domain/jira/model/JiraType";
 import {
     getAddableTypeName, getPinnableTypeName, getZeplinLinkTypeName, getJiraTypeName
