+++ src/common/domain/openInZeplin/command/OpenInZeplinCommand.ts
@@ -1,4 +1,4 @@
-import Command from "../../../common/vscode/command/Command";
+import Command from "../../../vscode/command/Command";
 import { openInZeplin } from "../flow/openInZeplinFlow";
 
 class OpenInZeplinCommand implements Command {
