+++ src/sidebar/activity/model/ComponentActivity.ts
@@ -3,7 +3,7 @@ import Activity from "./Activity";
 import { getComponentUri } from "../../../common/domain/openInZeplin/util/zeplinUris";
 import ApplicationType from "../../../common/domain/openInZeplin/model/ApplicationType";
 import BarrelType from "../../../common/domain/barrel/BarrelType";
-import ZeplinLinkType from "../../openInZeplin/model/ZeplinLinkType";
+import ZeplinLinkType from "../../../common/domain/openInZeplin/model/ZeplinLinkType";
 
 export default class ComponentActivity extends Activity {
     public constructor(private component: ZeplinComponent, private barrelId: string, private barrelType: BarrelType) {
