+++ src/sidebar/screen/tree/ScreenSectionTreeItem.ts
@@ -5,10 +5,10 @@ import ScreenTreeItem from "./ScreenTreeItem";
 import TreeItemContextProvider from "../../../common/vscode/tree/TreeItemContextProvider";
 import TreeItemContext from "../../../common/domain/tree/TreeItemContext";
 import Barrel from "../../../common/domain/barrel/Barrel";
-import ZeplinUriProvider from "../../openInZeplin/model/ZeplinUriProvider";
+import ZeplinUriProvider from "../../../common/domain/openInZeplin/model/ZeplinUriProvider";
 import ApplicationType from "../../../common/domain/openInZeplin/model/ApplicationType";
 import { getScreenSectionUri } from "../../../common/domain/openInZeplin/util/zeplinUris";
-import ZeplinLinkType from "../../openInZeplin/model/ZeplinLinkType";
+import ZeplinLinkType from "../../../common/domain/openInZeplin/model/ZeplinLinkType";
 
 function getContextProvider(section: ScreenSection): TreeItemContextProvider {
     const contexts = [TreeItemContext.ScreenSection, TreeItemContext.ZeplinLink];
